Use utils/getPostcssLoader in webpack.proto.js
Updates to the dev and prod webpack configs changed the way the postcss-loader is configured in ways that were not mirrored in webpack.proto.js. As a result, `yarn run proto` fails because it can't find a postcss.config.js file. This change fixes that issue (though there appear to be some other issues issues that break `yarn run proto`, as seen in issue #567).