add watchDirs to kyt config (#638)
`kyt dev`: for projects with a server, the only directory that is watched by `chokidar` is `src/server`. This is inadequate if the project relies on modules with alias'd paths outside of `src/server`.

Example:
projects/amp/src/server/index.js relies on
projects/components/src/Span/index.js

watchDirs is an optional flag that can be added so that `chokidar` reacts to the alias'd files changing and restarts the server.