fix: silence "loose" warnings from babel plugin(s)
It was brought to our attention by @ilyaGurevich (Thanks Ilya!) that
various warnings about inconsistant usage of the `loose` option for a
few babel plugins.

This commit sets the `loose` option explicitly for
`@babel/plugin-proposal-private-property-in-object` to silence the
warning, which @kohlmannj found was added to `@babel/preset-env` in
babel/babel#13176.

This should resolve the warnings folks had been seeing.

Closes #906