build!: rm `targets`, add .browserslistrc (#1003)
BREAKING CHANGE: `targets` are no longer set for `@babel/preset-env`.

In this commit, we introduced a breaking change in the way that browser
`targets` work, namely in that the targets are no longer set and the
user is expected to provide their own `.browserslistrc` or similar
configuration.

The previous behavior was to have browser `targets`, which overrides any
other configuration, including `.browserslistrc`.

From the `@babel/preset-env` documentation[^1]:
> By default `@babel/preset-env` will use browserslist config sources
> unless either the `targets` or `ignoreBrowserslistConfig` options are
> set.

When upgrading to the next major version that includes this change, *you
must provide your own configuration* or else Babel will default to using
the oldest browsers available[^2].

1. https://babeljs.io/docs/en/babel-preset-env#browserslist-integration
2. https://babeljs.io/docs/en/options#no-targets